Change data type from numeric to integer for balance column in the accounts table

+++ backend/db/migrations/20241127220930_create_accounts_table.sql
@@ -5,5 +5,5 @@ CREATE TABLE
   accounts (
     id SERIAL PRIMARY KEY,
     user_id INTEGER NOT NULL REFERENCES users (id) ON DELETE CASCADE,
-    balance NUMERIC(12, 2) NOT NULL
+    balance INTEGER NOT NULL
   )
